Issue 21766: Update where BCL URI validity is checked
